适配IOS，解决IOS10警告(master)

+++ 30000day/1.HomePage/1.MainPage/controller/SportsFunctionViewController.m
@@ -21,45 +21,27 @@
     [super viewDidLoad];
     
     SportsFunctionManager *sportsFunctionManager = [[SportsFunctionManager alloc] init];
-    
     self.sportsFunctionManager = sportsFunctionManager;
-    
-    
     [self.view setBackgroundColor:VIEWBORDERLINECOLOR];
     
-    
     UIView *view = [[UIView alloc] initWithFrame:CGRectMake(0, 0, SCREEN_WIDTH, 64)];
-    
     [view setBackgroundColor:[UIColor whiteColor]];
-    
     [self.view addSubview:view];
     
-    
     UILabel *lable = [[UILabel alloc] initWithFrame:CGRectMake((SCREEN_WIDTH - 80) / 2, 34, 80, 20)];
-    
     [lable setTextAlignment:NSTextAlignmentCenter];
-    
     if (self.type) [lable setText:@"地图设置"];else [lable setText:@"播报距离"];
-    
     [view addSubview:lable];
 
-    
     UITableView *tableView = [[UITableView alloc] initWithFrame:CGRectMake(0, 64.5, SCREEN_WIDTH, SCREEN_HEIGHT)];
-    
     [tableView setDataSource:self];
-    
     [tableView setDelegate:self];
-    
     tableView.tableFooterView = [[UIView alloc] init];
-    
     [self.view addSubview:tableView];
-    
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-
     return self.dataArray.count;
-
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
@@ -67,17 +49,12 @@
     UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:@"cell"];
     
     if (!cell) {
-        
         cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:@"cell"];
-        
     }
     
     cell.textLabel.text = self.dataArray[indexPath.row];
-    
     if (indexPath.row == self.selectIndex) {
-        
         [cell setAccessoryType:UITableViewCellAccessoryCheckmark];
-        
     }
     
     return cell;
@@ -87,28 +64,21 @@
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
     
     UITableViewCell *cell = [tableView cellForRowAtIndexPath:indexPath];
-    
     [cell setAccessoryType:UITableViewCellAccessoryCheckmark];
     
     if (self.type) {
 
         //修改数据
-        [self.sportsFunctionManager updateSportsFunction:STUserAccountHandler.userProfile.userId mapType:[NSString stringWithFormat:@"%ld",indexPath.row]];
-        
+        [self.sportsFunctionManager updateSportsFunction:STUserAccountHandler.userProfile.userId mapType:[NSString stringWithFormat:@"%ld",(long)indexPath.row]];
         [self dismissViewControllerAnimated:YES completion:^{
-            
             [STNotificationCenter postNotificationName:STSelectSportsFunctionMapSendNotification object:@(indexPath.row)];
-            
         }];
         
     } else {
     
-        [self.sportsFunctionManager updateSportsFunction:STUserAccountHandler.userProfile.userId speechDistance:[NSString stringWithFormat:@"%ld",indexPath.row]];
-        
+        [self.sportsFunctionManager updateSportsFunction:STUserAccountHandler.userProfile.userId speechDistance:[NSString stringWithFormat:@"%ld",(long)indexPath.row]];
         [self dismissViewControllerAnimated:YES completion:^{
-            
             [STNotificationCenter postNotificationName:STSelectSportsFunctionSpeechDistanceSendNotification object:@(indexPath.row)];
-            
         }];
     
     }
